+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Change Log
 
+* **master**
+    * Added `trackExistence` index option.
+    * Added `not` index method to support `Not` queries. The corresponding index must be created with `trackExistence=true` option.  
+
 * **v1.0.1** (2018-09-12)
     * Compatible with Pilosa 1.0.
     * Added key import support.