+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## Change Log
 
-* **master**
+* **v1.0.2** (2018-10-12)
     * Added `trackExistence` index option.
     * Added `not` index method to support `Not` queries. The corresponding index must be created with `trackExistence=true` option. This feature requires Pilosa on master branch. 
-    * Added support for roaring imports which can speed up the import process by %30 for non-key column imports. Pass `setRoaring(true)`` to `ImportOptions.builder()` to enable it. This feature requires Pilosa on master branch.   
+    * Added support for roaring imports which can speed up the import process by %30 for non-key column imports. Pass `setRoaring(true)`` to `ImportOptions.builder()` to enable it. This feature requires Pilosa on master branch.
+    * Fixes: `Column.create` method. See: https://github.com/pilosa/java-pilosa/pull/127    
 
 * **v1.0.1** (2018-09-12)
     * Compatible with Pilosa 1.0.